chore(refactor-react-components): move form components into folder and update paths in SignUp and SignIn form
